# docs/swagger.yaml
basePath: /
definitions:
...
  domain.GetRestaurant:
    properties:
      address:
        type: string
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      email:
        type: string
      id:
        type: string
      image:
        type: string
      isActive:
        type: boolean
      mediumPrice:
        type: number
      number:
        type: string
      rating:
        type: number
      timeWorkEnd:
        type: string
      timeWorkStart:
        type: string
      title:
        type: string
    type: object
  domain.GetRestaurantByAddress:
    properties:
      address:
        type: string
    type: object
  domain.GetRestaurantCategories:
    properties:
      category:
        type: string
      restaurants:
        items:
          $ref: '#/definitions/domain.GetRestaurant'
        type: array
    type: object
  domain.GetRestaurantOrderBy:
    properties:
      categories:
        type: string
      order_by:
        type: string
      sort_by:
        type: string
      title:
        type: string
    type: object
...
  domain.Restaurant:
    properties:
      address:
        type: string
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      email:
        type: string
      image:
        type: string
      number:
        type: string
      timeWorkEnd:
        type: string
      timeWorkStart:
        type: string
      title:
        type: string
      userId:
        type: integer
    required:
    - title
    type: object
...
  domain.UpdateRestaurant:
    properties:
      address:
        type: string
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      email:
        type: string
      image:
        type: string
      isActive:
        type: boolean
      number:
        type: string
      timeWorkEnd:
        type: string
      timeWorkStart:
        type: string
      title:
        type: string
    type: object
...
  handler.dataResponse:
    properties:
      data: {}
    type: object
...
  handler.getAllRestaurantsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GetRestaurant'
        type: array
    type: object
  handler.getCategories:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
...
  handler.getRestaurant:
    properties:
      data:
        $ref: '#/definitions/domain.GetRestaurant'
    type: object
  handler.getRestaurantsCategoriesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GetRestaurantCategories'
        type: array
    type: object
  handler.response:
    properties:
      message:
        type: string
    type: object
...
host: 165.232.68.67:8080
info:
  contact: {}
  description: REST API for Restaurant-Assistant service
  title: Restaurant-Assistant API
  version: "1.0"
paths:
...
  /restaurant:
    post:
      consumes:
      - application/json
      description: create restaurant
      parameters:
      - description: create restaurant
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Restaurant'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
  /restaurant/:
    get:
      description: getting restaurants
      parameters:
      - description: sort restaurant by rating and title
        in: body
        name: sort_by
        schema:
          $ref: '#/definitions/domain.GetRestaurantOrderBy'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getAllRestaurantsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
  /restaurant/{id}:
    delete:
      description: delete restaurant
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
    get:
      description: getting restaurant by id
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getRestaurant'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
    put:
      consumes:
      - application/json
      description: update restaurant
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: create restaurant
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateRestaurant'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getRestaurant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
...
  /restaurant/categories:
    get:
      description: getting restaurant categories
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getCategories'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
  /restaurant/categories-with-restaurants:
    get:
      description: getting restaurant categories with restaurants
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getRestaurantsCategoriesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
...
  /restaurant/nearest:
    get:
      description: getting restaurants
      parameters:
      - description: get nearest restaurants
        in: body
        name: address
        schema:
          $ref: '#/definitions/domain.GetRestaurantByAddress'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/handler.getAllRestaurantsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.response'
      tags:
      - restaurant
securityDefinitions:
  ManagerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"